requirements.txt
mysql-connector-python==8.0.33
Pillow==10.4.0
pygame==2.5.2
python-chess==1.0.0
werkzeug==2.3.6